requirements.txt
# ── FILE: requirements.txt ──
# AI-Based Fake News Detection - Python Dependencies

flask==3.0.0
scikit-learn==1.6.1
nltk==3.9.1
pandas==2.2.3
numpy==2.2.3
joblib==1.4.2
requests==2.32.3
python-dotenv==1.0.1
